+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ## Added
 - Added [BaryScore, InfoLM, and DepthScore](models/colombo2021/Readme.md)
+- Added ability to set `beam_size` and `nbest` parameters for BART.
 
 ### Changed
 - Changed the backend implementation of MoverScore to use a non-IDF dict based version.